Add: HTML templates for latest photo endpoint plus code refactor
