+++ app/routes/index.js
@@ -5,7 +5,7 @@ const router = express.Router();
  * Default Page
  */
 router.get('/', (req, res) => {
-    res.status(200).send('Welcome to NASA Simple API Integration');
+    res.status(200).render('../views/indexView.html')
 });
 
 module.exports = router;
