+++ app/views/meteorsView.html
@@ -2,10 +2,10 @@
 <html lang="en">
     <head>
         <title>Meteors Page</title>
-        <h2>Meteors registered From Date: {{dateFrom}}  To Date: {{dateTo}}</h2>
+        <h2>Meteors registered From Date: {{dateFrom}} To Date: {{dateTo}}</h2>
         <style>
-            div {margin: auto; width: 50%; text-align: center; padding-top: 20px}
-            h2 {margin: auto; width: 50%; text-align: center;}
+            div {margin: auto; text-align: center; padding-top: 20px}
+            h2 {margin: auto; text-align: center;}
         </style>
     </head>
     <body>
