+++ app/views/photoView.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <title>Mars Photo Page</title>
+        <h2>Latest photo</h2>
+        <style>
+            div {margin: auto; text-align: center;}
+            h2 {margin: auto; width: 50%; text-align: center;}
+        </style>
+    </head>
+    <body>
+        <div>
+            <img src='{{latestPhoto}}' alt="Latest Photo">
+        </div>
+    </body>
+</html>
